Create EXP 1
Its sliding puzzle, it consists of a 3x3 grid with eight numbered tiles 
and one empty space. The goal of the puzzle is to rearrange the tiles 
from their initial, scrambled state to a goal state where the numbers 
are ordered from 1 to 8, with the empty space in the bottom-right 
corner. Implemented using python