Create EXP 5
It's a river-crossing puzzle that involves three missionaries and 
three cannibals who need to cross a river from one side to another. 
The problem comes with certain constraints: 
 There is a boat that can carry at most two people at a time. 
 On either side of the river, if the number of cannibals ever 
exceeds the number of missionaries, the cannibals will eat 
the missionaries, and the mission will fail. The goal is to find a series of moves that allow all the missionaries and cannibals to cross the river safely. Implement the scenario using Python